Add files via upload
the projected was submitted on March 12th, 2024, at 11:46 am
I Uploaded this document 2nd time. because the Google Dive link opens the PowerPoint presentation in Google Slides, due to this the layout is rearranged. I have submitted the project in time. there is this small bug. so I am updating this doc.

The first document was uploaded due to the PowerPoint file size. GitHub has rules the file should be under 25MB, so to share the PowerPoint and a website link I added it in the docs.

for quick access:
Link to Powe point: 
https://1drv.ms/p/s!AgxHClk6OkmXgRe48ysyEFoaRT5a?e=5h49uL
Link to website:
https://arivazhagan1999.my.canva.site/capstone-project-1